refactor(util) append _service suffix to all service files
